% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feglm_ml_wrappers.R
\name{benchmark_ml}
\alias{benchmark_ml}
\title{Performance comparison between original and ML implementations}
\usage{
benchmark_ml(formula, data, family = poisson(), weights = NULL, n_runs = 5)
}
\arguments{
\item{formula}{an object of class \code{"formula"}}

\item{data}{an object of class \code{"data.frame"}}

\item{family}{GLM family to test}

\item{weights}{optional weights}

\item{n_runs}{number of benchmark runs}
}
\value{
A list with timing and memory comparisons
}
\description{
Compare speed and memory usage between capybara and capybara_ml
}
