% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binarization.R
\name{binarization.mann.whitney}
\alias{binarization.mann.whitney}
\title{Binarization and Identity Calling from Identity Score}
\usage{
binarization.mann.whitney(mtx, ref.perc.ls, ref.meta, perc.ls,
  bulk = FALSE, map.df = NULL)
}
\arguments{
\item{mtx}{The matrix of identity scores of query cells. The row number is the total number of query cells and the column number is the number of total number of possible cell types}

\item{ref.perc.ls}{Emprical p-values for reference cells}

\item{ref.meta}{T}

\item{perc.ls}{Emprical p-values for query cells}

\item{bulk}{If the reference data type is bulk RNA-seq. The default is bulk = FALSE}

\item{map.df}{bulk mapping. The default is bulk = FALSE}
}
\description{
This function calls single identity or multiple identities of query cells from empirical p-values. Inferred cell types are marked by 1 in a binarized matrix.
}
\examples{

}
\keyword{binarization,}
\keyword{calling}
\keyword{identity}
