% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification.R
\name{binary.to.classification}
\alias{binary.to.classification}
\title{Generate classification results from the binarized result matrix}
\usage{
binary.to.classification(bin.count.rslt)
}
\arguments{
\item{bin.count.rslt}{The binarized classification result matrix}
}
\description{
This function genrate classification table from the binarized result matrix
}
\note{

}
\examples{

}
\keyword{calssification,}
\keyword{result}
