% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_augment.R, R/generics_glance.R,
%   R/generics_tidy.R
\name{augment.feglm}
\alias{augment.feglm}
\alias{augment.felm}
\alias{glance.feglm}
\alias{glance.felm}
\alias{tidy.feglm}
\alias{tidy.felm}
\title{Broom Integration}
\usage{
\method{augment}{feglm}(x, newdata = NULL, ...)

\method{augment}{felm}(x, newdata = NULL, ...)

\method{glance}{feglm}(x, ...)

\method{glance}{felm}(x, ...)

\method{tidy}{feglm}(x, conf_int = FALSE, conf_level = 0.95, ...)

\method{tidy}{felm}(x, conf_int = FALSE, conf_level = 0.95, ...)
}
\arguments{
\item{x}{A fitted model object.}

\item{newdata}{Optional argument to use data different from the data used to fit the model.}

\item{...}{Additional arguments passed to the method.}

\item{conf_int}{Logical indicating whether to include the confidence interval.}

\item{conf_level}{The confidence level for the confidence interval.}
}
\value{
A tibble with the respective information for the \code{augment}, \code{glance}, and \code{tidy} methods.
}
\description{
The provided \code{broom} methods do the following:
\enumerate{
\item \code{augment}: Takes the input data and adds additional columns with the fitted values and residuals.
\item \code{glance}: Extracts the deviance, null deviance, and the number of observations.`
\item \code{tidy}: Extracts the estimated coefficients and their standard errors.
}
}
\examples{
ross2004_subset <- ross2004[ross2004$year == 1999, ]
ross2004_subset <- ross2004_subset[ross2004_subset$ltrade > 0, ]

fit <- fepoisson(ltrade ~ ldist, ross2004_subset, control = fit_control(keep_data = TRUE))

broom::augment(fit)
broom::glance(fit)
broom::tidy(fit)

}
