% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellQuadraticProgrammingAux.R
\name{calc.scale.ratio}
\alias{calc.scale.ratio}
\title{Scale Ratio Calculation}
\usage{
calc.scale.ratio(bulk, sc)
}
\arguments{
\item{bulk}{The reference dataset}

\item{sc}{The single-cell dataset}
}
\description{
This function calculate the scale ratio between the single-cell transcriptome and reference dataset to minimize the Lagrangian multiplier. In the other word, reduce the restriction.
}
\examples{
calc.scale.ratio(reference.dt, single.cell.mtx)
}
\keyword{Scale}
\keyword{calculation}
\keyword{ratio}
