% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feglm_ml.R
\name{center_variables_ml}
\alias{center_variables_ml}
\title{Fast ML-based centering of variables}
\usage{
center_variables_ml(V, w, k_list, tol = 1e-06, max_iter = 10000L)
}
\arguments{
\item{V}{matrix or data frame to center}

\item{w}{weights vector}

\item{k_list}{list of fixed effects indices}

\item{tol}{convergence tolerance}

\item{max_iter}{maximum iterations}
}
\value{
Centered matrix
}
\description{
Maximum likelihood approach to centering variables
}
