% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in generics_coef.R
\name{coef.capybara_model}
\alias{coef.capybara_model}
\title{Extract coefficients from a 'capybara_model' object}
\usage{
\method{coef}{capybara_model}(object, ...)
}

\arguments{
\item{object}{a 'capybara_model' object}

\item{...}{additional arguments for S3 compliance (unused)}
}

\description{
Shared by 'feglm' and 'felm' objects (and their subclasses, e.g.
 'fenegbin', 'fepoisson_asymmetric'), similar to the 'coef' method for 'glm'/'lm' objects.
}

