% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in generics_coef.R
\name{coef.felm}
\alias{coef.felm}
\title{Extract coefficients from 'felm' object}
\usage{
\method{coef}{felm}(object, ...)
}

\arguments{
\item{object}{'felm' object}

\item{...}{additional arguments for S3 compliance (unused)}
}

\description{
Similar to the 'coef' method for 'lm' objects.
}

