% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in generics_coef.R
\name{coef.summary.feglm}
\alias{coef.summary.feglm}
\title{Extract coefficients from 'summary.feglm' object}
\usage{
\method{coef}{summary.feglm}(object, ...)
}

\arguments{
\item{object}{'summary.feglm' object}

\item{...}{additional arguments for S3 compliance (unused)}
}

\description{
Similar to the 'coef' method for 'summary.glm' objects.
}

