% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in generics_coef.R
\name{coef.summary.felm}
\alias{coef.summary.felm}
\title{Extract coefficients from 'summary.felm' object}
\usage{
\method{coef}{summary.felm}(object, ...)
}

\arguments{
\item{object}{'summary.felm' object}

\item{...}{additional arguments for S3 compliance (unused)}
}

\description{
Similar to the 'coef' method for 'summary.felm' objects.
}

