% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighResolutionReferenceConstruction.R
\name{con.high.res.reference}
\alias{con.high.res.reference}
\title{Systematic construction of a high-resolution reference}
\usage{
con.high.res.reference(ref.mtx, coldata.df, cell.num.for.ref = 90)
}
\arguments{
\item{ref.mtx}{The single-cell reference dataset}

\item{coldata.df}{The metadata (cell type information) for cells in the high-resolution reference}

\item{cell.num.for.ref}{The number of cell numbers used to build the reference for each cell type. The default is cell.num.for.ref = 90}
}
\description{
This function returns a
}
\examples{

}
\keyword{pseudo-bulk}
\keyword{reference}
