% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighResolutionReferenceConstruction.R
\name{construct.high.res.reference}
\alias{construct.high.res.reference}
\title{Systematic construction of a high-resolution reference}
\usage{
construct.high.res.reference(ref.mtx, coldata.df, criteria,
  cell.num.for.ref = 90)
}
\arguments{
\item{ref.mtx}{The single-cell reference dataset}

\item{coldata.df}{The metadata (cell type information) for cells in the high-resolution reference}

\item{cell.num.for.ref}{The number of cell numbers used to build the reference for each cell type. The default is cell.num.for.ref = 90}
}
\description{
Create a pseudo-bulk reference by sampling 90-cells from each cell type to maintain cellular resolution while increasing transcriptional resolution
}
\examples{

}
\keyword{pseudo-bulk}
\keyword{reference}
