% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capybara-package.R
\docType{data}
\name{correia2019}
\alias{correia2019}
\title{Separation Example Datasets}
\format{
\subsection{\code{correia2019}}{

A list of data frames with three elements:
\describe{
\item{example1}{Data frame used to show lack of convergence}
\item{example2}{Data frame used to show lack of convergence with step-halving}
\item{fe1}{Data frame with fixed effects used with the 'alpaca' package}
}
}
}
\source{
'correia2019' GitHub repository (https://github.com/sergiocorreia/correia2019/blob/master/guides/nonexistence_benchmarks.md#r-packages)
}
\usage{
correia2019
}
\description{
Nonexistence of estimates of Poisson models across different statistical packages.
}
\keyword{datasets}
