% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feglm_ml.R
\name{feglm_ml}
\alias{feglm_ml}
\title{Maximum Likelihood GLM fitting with high-dimensional k-way fixed effects}
\usage{
feglm_ml(
  formula = NULL,
  data = NULL,
  family = poisson(),
  weights = NULL,
  beta_start = NULL,
  eta_start = NULL,
  control = NULL,
  ...
)
}
\arguments{
\item{formula}{an object of class \code{"formula"}}

\item{data}{an object of class \code{"data.frame"}}

\item{family}{a GLM family}

\item{weights}{an optional vector of weights}

\item{beta_start}{optional starting values for coefficients}

\item{eta_start}{optional starting values for linear predictors}

\item{control}{a list of control parameters}

\item{...}{additional arguments}
}
\value{
A named list of class \code{"feglm_ml"}
}
\description{
Fast ML-based implementation following FENmlm approach
}
