% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feglm_ml.R
\name{felm_ml}
\alias{felm_ml}
\title{Maximum Likelihood Linear Model fitting with high-dimensional k-way fixed effects}
\usage{
felm_ml(formula = NULL, data = NULL, weights = NULL, control = NULL, ...)
}
\arguments{
\item{formula}{an object of class \code{"formula"}}

\item{data}{an object of class \code{"data.frame"}}

\item{weights}{an optional vector of weights}

\item{control}{a list of control parameters}

\item{...}{additional arguments}
}
\value{
A named list of class \code{"felm_ml"}
}
\description{
Fast ML-based linear model implementation
}
