% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feglm_ml_wrappers.R
\name{felogit_ml}
\alias{felogit_ml}
\title{Fast ML-based Logistic regression with fixed effects}
\usage{
felogit_ml(formula = NULL, data = NULL, weights = NULL, control = NULL, ...)
}
\arguments{
\item{formula}{an object of class \code{"formula"}}

\item{data}{an object of class \code{"data.frame"}}

\item{weights}{an optional vector of weights}

\item{control}{a list of control parameters}

\item{...}{additional arguments}
}
\value{
A named list of class \code{"feglm_ml"}
}
\description{
Optimized Logistic GLM using maximum likelihood approach
}
