% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feglm_ml_wrappers.R
\name{fenegbin_ml}
\alias{fenegbin_ml}
\title{Fast ML-based Negative Binomial regression with fixed effects}
\usage{
fenegbin_ml(
  formula = NULL,
  data = NULL,
  weights = NULL,
  theta = 1,
  control = NULL,
  ...
)
}
\arguments{
\item{formula}{an object of class \code{"formula"}}

\item{data}{an object of class \code{"data.frame"}}

\item{weights}{an optional vector of weights}

\item{theta}{dispersion parameter}

\item{control}{a list of control parameters}

\item{...}{additional arguments}
}
\value{
A named list of class \code{"feglm_ml"}
}
\description{
Optimized Negative Binomial GLM using maximum likelihood approach
}
