% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fepoisson_asymmetric.R
\name{fepoisson_asymmetric}
\alias{fepoisson_asymmetric}
\title{Asymmetric Poisson Pseudo-Maximum Likelihood (APPML) Estimation}
\usage{
fepoisson_asymmetric(
  formula = NULL,
  data = NULL,
  weights = NULL,
  beta_start = NULL,
  eta_start = NULL,
  offset = NULL,
  control = NULL
)
}
\arguments{
\item{formula}{an object of class \code{"formula"}: a symbolic description of the model to be fitted. \code{formula}
must be of type \code{response ~ slopes | fixed_effects | cluster}.}

\item{data}{an object of class \code{"data.frame"} containing the variables in the model. The expected input is a
dataset with the variables specified in \code{formula} and a number of rows at least equal to the number of variables
in the model.}

\item{weights}{an optional string with the name of the prior weights variable in \code{data}.}

\item{beta_start}{an optional vector of starting values for the structural parameters in the linear predictor.
Default is \eqn{\boldsymbol{\beta} = \mathbf{0}}{\beta = 0}.}

\item{eta_start}{an optional vector of starting values for the linear predictor.}

\item{offset}{an optional formula or numeric vector specifying an a priori known component to be included in the
linear predictor. If a formula, it should be of the form \code{~ variable}.}

\item{control}{a named list of parameters for controlling the fitting process. See \link{fit_control} for details.}
}
\value{
A named list of class \code{"feglm"} containing:
\item{coefficients}{named vector of estimated coefficients}
\item{vcov}{variance-covariance matrix of coefficients}
\item{eta}{linear predictor}
\item{fitted_values}{fitted values from the final iteration}
\item{residuals}{residuals from the final fit}
\item{weights}{observation weights used in final fit}
\item{appml_weights}{asymmetric weights used in APPML algorithm}
\item{deviance}{the deviance of the model}
\item{null_deviance}{the null deviance of the model}
\item{conv}{logical indicating whether inner GLM converged}
\item{conv_outer}{logical indicating whether APPML outer loop converged}
\item{iter}{number of inner iterations}
\item{iter_outer}{number of outer APPML iterations}
\item{expectile}{the expectile value used}
\item{objective_function}{final value of the convergence criterion}
\item{negative_residuals_share}{proportion of negative residuals in final fit}
\item{nobs}{a named vector with the number of observations}
\item{fe_levels}{a named vector with the number of levels in each fixed effect}
\item{nms_fe}{a list with the names of the fixed effects variables}
\item{formula}{the formula used in the model}
\item{family}{the family used in the model (Poisson)}
\item{control}{the control list used in the model}
}
\description{
Fits an asymmetric Poisson pseudo-maximum likelihood model with high-dimensional fixed effects
using expectile regression. This approach extends standard PPML by allowing different weights for positive
and negative residuals, enabling estimation of conditional expectiles rather than the conditional mean.
}
\details{
The APPML estimator minimizes an asymmetric loss function based on expectiles. For a given expectile \eqn{\tau},
observations with negative residuals receive weight \eqn{\tau} while observations with positive residuals
receive weight \eqn{1 - \tau}. The algorithm iteratively:
\enumerate{
\item Computes residuals from the current fit
\item Updates weights as \eqn{w_i = |\tau - \mathbf{1}(r_i < 0)|}
\item Re-fits the weighted Poisson model
\item Checks convergence using \eqn{(b - b_{old})' V^{-1} (b - b_{old}) < \epsilon}
}

The expectile parameter is specified via \code{control = fit_control(expectile = ...)}. When
\code{expectile = 0.5}, the estimator is equivalent to standard PPML. Values below 0.5 estimate
lower conditional expectiles (more sensitive to small values), while values above 0.5 estimate
upper conditional expectiles (more sensitive to large values).
}
\examples{
ross2004_subset <- ross2004[ross2004$year == 1999, ]
ross2004_subset <- ross2004_subset[ross2004_subset$ltrade > 0, ]

# Lower expectile (10th) - more weight on negative residuals
fit10 <- fepoisson_asymmetric(
  ltrade ~ ldist | ctry1, ross2004_subset,
  control = fit_control(expectile = 0.1)
)

summary(fit10)

}
\references{
Newey, W. K., & Powell, J. L. (1987). Asymmetric least squares estimation and testing.
\emph{Econometrica}, 55(4), 819-847.
}
\seealso{
\link{fepoisson}, \link{feglm}, \link{fit_control}
}
