% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feprobit.R
\name{feprobit}
\alias{feprobit}
\title{Probit model fitting with high-dimensional k-way fixed effects}
\usage{
feprobit(
  formula = NULL,
  data = NULL,
  weights = NULL,
  vcov = NULL,
  beta_start = NULL,
  eta_start = NULL,
  offset = NULL,
  control = NULL
)
}
\arguments{
\item{formula}{an object of class \code{"formula"}: a symbolic description of the model to be fitted. \code{formula}
must be of type \code{response ~ slopes | fixed_effects | cluster}.}

\item{data}{an object of class \code{"data.frame"} containing the variables in the model. The expected input is a
dataset with the variables specified in \code{formula} and a number of rows at least equal to the number of variables
in the model.}

\item{weights}{an optional string with the name of the prior weights variable in \code{data}.}

\item{vcov}{an optional character string specifying the type of variance-covariance estimator.
One of \code{"iid"} (default OLS, ignore cluster part of formula), \code{"hetero"} (heteroskedastic-robust
HC0, computed in C++ - no cluster variable needed), \code{"cluster"} (one-way sandwich using the cluster
variable in the formula), \code{"m-estimator"} (M-estimator one-way sandwich), or \code{"dyadic"}
(Cameron-Miller dyadic sandwich; requires two entity variables in the third part of the formula).
When \code{NULL} (default), the type is inferred from the formula: if a cluster variable is present the
standard sandwich is used, otherwise the inverse Hessian (IID) is returned.}

\item{beta_start}{an optional vector of starting values for the structural parameters in the linear predictor.
Default is \eqn{\boldsymbol{\beta} = \mathbf{0}}{\beta = 0}.}

\item{eta_start}{an optional vector of starting values for the linear predictor.}

\item{offset}{an optional formula or numeric vector specifying an a priori known component to be included in the
linear predictor. If a formula, it should be of the form \code{~ variable}.}

\item{control}{a named list of parameters for controlling the fitting process. See \link{fit_control} for details.}
}
\value{
A named list of class \code{"feglm"}.
}
\description{
A wrapper for \link{feglm} with
\code{family = probit()}. This fits a binary response model using
the probit link function (standard normal CDF).
}
\examples{
# check the feglm examples for details about clustered standard errors
mod <- feprobit(am ~ wt | cyl, mtcars)
summary(mod)

}
