% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in generics_fitted.R
\name{fitted.feglm}
\alias{fitted.feglm}
\title{Extract fitted values from 'feglm' object}
\usage{
\method{fitted}{feglm}(object, ...)
}

\arguments{
\item{object}{'feglm' object}

\item{...}{additional arguments for S3 compliance (unused)}
}

\description{
Similar to the 'fitted' method for 'glm' objects.
}

