% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in generics_fitted.R
\name{fitted.felm}
\alias{fitted.felm}
\title{Extract fitted values from 'felm' object}
\usage{
\method{fitted}{felm}(object, ...)
}

\arguments{
\item{object}{'felm' object}

\item{...}{additional arguments for S3 compliance (unused)}
}

\description{
Similar to the 'fitted' method for 'lm' objects.
}

