% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellQuadraticProgrammingAux.R
\name{gene.intersect.sub}
\alias{gene.intersect.sub}
\title{Find intersection genes}
\usage{
gene.intersect.sub(bulk, sc)
}
\arguments{
\item{bulk}{The reference dataset}

\item{sc}{The single-cell dataset}
}
\description{
This function identifies the intersection genes between reference and single-cell data
}
\examples{
gene.intersect.sub(reference.dt, single.cell.mtx)
}
