% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{get.lest.connected}
\alias{get.lest.connected}
\title{Get the Least Connected Cells}
\usage{
get.lest.connected(mtx, n.sample)
}
\arguments{
\item{mtx}{The normalized reference count matrix}

\item{n.sample}{The number of reference cells to be included within each pseudo-bulk}
}
\description{
This function returns index of cells that are the least connected within a cell type
}
