% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{get.mid.connected}
\alias{get.mid.connected}
\title{Get the Medium Connected Cells}
\usage{
get.mid.connected(mtx, n.sample)
}
\arguments{
\item{mtx}{The normalized reference count matrix}

\item{n.sample}{The number of reference cells to be included within each pseudo-bulk}
}
\description{
This function returns index of cells that are the most connected within a cell type
}
