% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellQuadraticProgrammingAux.R
\name{normalize.dt}
\alias{normalize.dt}
\title{Normalization of Single-Cell RNA-Seq Data}
\usage{
normalize.dt(dt.st)
}
\arguments{
\item{dt.st}{The dataset to normalize}
}
\description{
This function normalizes single-cell RNA-seq data using its raw counts. The normalization is performed to remove variation due to difference in read depth and coverage.
}
\examples{
normalize.dt(single.cell.mtx)
}
\keyword{normalization}
