% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmpiricalPValueCalculation.R
\name{perc.calc.aux}
\alias{perc.calc.aux}
\title{Emprical p-values for a query cell}
\usage{
perc.calc.aux(
  l,
  cell.names,
  mtx.ct.num,
  background.mtx,
  mtx,
  cnms,
  test.times = 50
)
}
\arguments{
\item{l}{The index of the sample cell of query}

\item{cell.names}{A vector of sample cells' names}

\item{mtx.ct.num}{The number of all possible cell types}

\item{background.mtx}{The matrix of identity score of reference cells. The row number is the total number of reference cells and the column number is the number of total number of possible cell types}

\item{mtx}{The matrix of identity scores of query cells. The row number is the total number of query cells and the column number is the number of total number of possible cell types}

\item{cnms}{The names of all possible cell types}

\item{test.times}{The number of times resampling the background. The default is test.times=50}
}
\description{
This function returns emprical p-values of all possilbe celltypes for a query cells under test.times resampled backgrounds.
}
\note{

}
\examples{

}
\keyword{empirical}
\keyword{p-value,}
\keyword{query}
\keyword{resampling,}
\keyword{single}
