% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmpiricalPValueCalculation.R
\name{percentage.calc}
\alias{percentage.calc}
\title{Emprical p-values for query cells}
\usage{
percentage.calc(mtx, bkgd.mtx)
}
\arguments{
\item{mtx}{The matrix of identity scores of query cells. The row number is the total number of query cells and the column number is the number of total number of possible cell types}

\item{bkgd.mtx}{The matrix of identity score of reference cells. The row number is the total number of reference cells and the column number is the number of total number of possible cell types}
}
\description{
This function returns a list of emprical p-value matrices. Every matrix in the list contains emprical p-values of all possilbe celltypes for a single query cells under test.times resampled backgrounds.
}
\note{

}
\examples{

}
\keyword{all}
\keyword{cells}
\keyword{empirical}
\keyword{p-value,}
\keyword{query}
\keyword{resampling,}
