% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in generics_plot.R
\name{plot.feglm}
\alias{plot.feglm}
\title{Plot method for feglm objects}
\usage{
\method{plot}{feglm}(x, ...)
}

\arguments{
\item{x}{A fitted model object of class \code{feglm}.}

\item{...}{Additional arguments passed to the method. In this case, the additional argument is `conf_level`, which is
 the confidence level for the confidence interval.}
}

\value{
No return value, called for the side effect of producing a plot
 of the estimated coefficients and their confidence intervals.
}

\description{
Plots the estimated coefficients and their confidence
 intervals.
}

\examples{
ross2004_subset <- ross2004[ross2004$year == 1999, ]
ross2004_subset <- ross2004_subset[ross2004_subset$ltrade >
  quantile(ross2004_subset$ltrade, 0.75), ]

fit <- fepoisson(ltrade ~ ldist | ctry1, ross2004_subset)

plot(fit, conf_level = 0.99)
}

