% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capybara-package.R
\docType{data}
\name{ppmlhdfe}
\alias{ppmlhdfe}
\title{'ppmlhdfe' separation example datasets}
\format{
\subsection{\code{ppmlhdfe}}{

A list of data frames with three elements:
\describe{
\item{example1}{Data frame used to show lack of convergence}
\item{example2}{Data frame used to show lack of convergence with step-halving}
\item{fe1}{Data frame with fixed effects used with the 'alpaca' package}
}
}
}
\source{
'ppmlhdfe' GitHub repository (https://github.com/sergiocorreia/ppmlhdfe/blob/master/guides/nonexistence_benchmarks.md#r-packages)
}
\usage{
ppmlhdfe
}
\description{
Nonexistence of estimates of Poisson models across different statistical packages.
}
\keyword{datasets}
