% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in generics_predict.R
\name{predict.feglm}
\alias{predict.feglm}
\title{Predict method for 'feglm' objects}
\usage{
\method{predict}{feglm}(object, newdata = NULL, type = c("response", "link"), ...)
}

\arguments{
\item{object}{'feglm' object}

\item{newdata}{'data.frame' to obtain new predictions}

\item{type}{'response' or 'link'}

\item{...}{additional arguments for S3 compliance (unused)}
}

\description{
Similar to the 'predict' method for 'glm' objects but returns response predictions as default.
}

