% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in generics_predict.R
\name{predict.felm}
\alias{predict.felm}
\title{Predict method for 'felm' objects}
\usage{
\method{predict}{felm}(object, newdata = NULL, type = c("response", "terms"), ...)
}

\arguments{
\item{object}{'felm' object}

\item{newdata}{'data.frame' to obtain new predictions}

\item{type}{'response' or 'terms'}

\item{...}{additional arguments for S3 compliance (unused)}
}

\description{
Similar to the 'predict' method for 'lm' objects
}

