% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in generics_print.R
\name{print.capybara_separation}
\alias{print.capybara_separation}
\title{Print method for regression tables}
\usage{
\method{print}{capybara_separation}(x, ...)
}

\arguments{
\item{x}{'feglm' object}

\item{...}{additional arguments for S3 compliance (unused)}
}

\description{
Print method for regression tables
}

