% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in generics_print.R
\name{print.felm}
\alias{print.felm}
\title{Print method for 'felm' objects (detailed output)}
\usage{
\method{print}{felm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}

\arguments{
\item{x}{'felm' object}

\item{digits}{how many digits to display}

\item{...}{additional arguments for S3 compliance (unused)}
}

\description{
Similar to the 'print' method for 'lm' objects
}

