% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in generics_print.R
\name{print.summary.felm}
\alias{print.summary.felm}
\title{Print method for 'summary.felm' objects}
\usage{
\method{print}{summary.felm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}

\arguments{
\item{x}{'summary.felm' object}

\item{digits}{how many digits to display}

\item{...}{additional arguments for S3 compliance (unused)}
}

\description{
Similar to the 'print' method for 'summary.lm' objects
}

