% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in generics_print.R
\name{print.summary_table}
\alias{print.summary_table}
\title{title Print method for regression tables}
\usage{
\method{print}{summary_table}(x, ...)
}

\arguments{
\item{x}{A summary_table object}

\item{...}{additional arguments for S3 compliance (unused)}
}

\description{
title Print method for regression tables
}

