% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in capybara-package.R
\name{ross2004}
\alias{ross2004}
\title{Subset of WTO data from Ross (2004)}
\description{
Data from Ross (2004) used to show the different variance-covariance estimators from
Cameron and Miller (2014).
}

