% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellQuadraticProgramming.R
\name{sc.quad.prog.run}
\alias{sc.quad.prog.run}
\title{Single-Cell Quadratic Programming Calculation}
\usage{
sc.quad.prog.run(
  bulk.transcriptome,
  single.cell.transcriptome,
  force.eq = 0,
  unix.parallel = FALSE,
  windows.parallel = FALSE,
  parallel.cores = 4
)
}
\arguments{
\item{bulk.transcriptome}{The reference transcriptome taht contains the transcriptome of each potential cell type}

\item{single.cell.transcriptome}{the transcriptome profile of cells from single-cell RNA-sequencing}

\item{force.eq}{either 0 or 1. Setting to 0 assumes the 1st constraint as inequality. Setting to 1 assumes equality. Default to 0}

\item{unix.parallel}{boolean value, either TRUE or FALSE. If using unix/linux based systems, this command can be set to TRUE to parallelize use parallel package. Default to FALSE}

\item{windows.parallel}{boolean value, either TRUE or FALSE. If using Windows based systems, this command can be set to TRUE to parallelize use snow package. Default to FALSE}

\item{parallel.cores}{the number of cores to use for parallel processes. If no parallelization selected, no parallelization will be implemented. Only 1 core will be used}
}
\description{
This function runs quadratic programming to identify the probability of the cells in single-cell RNA-seq belonging to cell types in the reference transcriptome
}
\note{
Code reference from Treutlein et. al., Dissecting direct reprogramming from fibroblast to neuron using single-cell RNA-seq
}
\examples{
sc.quad.prog.run(ref.transcriptome, sc.transcriptome, force.eq = 1)
}
\keyword{programming,}
\keyword{quadratic}
\keyword{scRNA-seq}
