% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellQuadraticProgramming.R
\name{single.round.QP.analysis}
\alias{single.round.QP.analysis}
\title{Single Round of Quadratic Programming}
\usage{
single.round.QP.analysis(
  ref,
  sc.data,
  scale.bulk.sc = "scale",
  unix.par = FALSE,
  windows.par = FALSE,
  force.eq = 0,
  n.cores = 1,
  save.to.path,
  save.to.filename,
  bulk.norm = T,
  norm.sc = T,
  log.bulk = T,
  log.sc = T
)
}
\arguments{
\item{ref}{The reference transcriptome taht contains the transcriptome of each potential cell type}

\item{sc.data}{the transcriptome profile of cells from single-cell RNA-sequencing}

\item{scale.bulk.sc}{either scale or non-scale. Scaling is recommended to make the reference comparable to the single-cell. Default to scale.}

\item{unix.par}{boolean value, either TRUE or FALSE. If using unix/linux based systems, this command can be set to TRUE to parallelize use parallel package. Default to FALSE}

\item{windows.par}{boolean value, either TRUE or FALSE. If using Windows based systems, this command can be set to TRUE to parallelize use snow package. Default to FALSE}

\item{force.eq}{either 0 or 1. Setting to 0 assumes the 1st constraint as inequality. Setting to 1 assumes equality. Default to 0}

\item{n.cores}{the number of cores to use for parallel processes. If no parallelization selected, no parallelization will be implemented. Only 1 core will be used}

\item{save.to.path}{which directory would you like to save your file?}

\item{save.to.filename}{prefix to the filename to save to. The final filename will be constructed inside of the function by tagging the following string _scale.csv for scaling and _non_scale.csv for not scaling.}

\item{bulk.norm}{boolean value, either TRUE or FALSE. Would you like to normalize the reference?}

\item{norm.sc}{boolean value, either TRUE or FALSE. Would you like to normalize the single-cell dataset?}

\item{log.bulk}{boolean value, either TRUE or FALSE. Would you like to log the reference?}

\item{log.sc}{boolean value, either TRUE or FALSE. Would you like to log the single-cell dataset?}
}
\description{
This function runs preprocessing, including log normalization, gene intersection and scaling. Further, run one round of quadratic programming. See function sc.quad.prog.run for detailed description
}
\note{
This function calls the quadratic programming referenced to Treutlein et. al.
}
\examples{
single.round.QP.analysis(ref = ref.transcriptome, sc.data = sc.transcriptome, force.eq = 1, save.to.path = "~/Desktop/", save.to.filename = "my_favorite")
}
\keyword{programming,}
\keyword{quadratic}
\keyword{scRNA-seq,}
