% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_summary.R
\name{summary.apes}
\alias{summary.apes}
\title{Summary method for fixed effects APEs}
\usage{
\method{summary}{apes}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"apes"}.}

\item{...}{additional arguments.}
}
\value{
A named matrix of covariance estimates.
}
\description{
Covariance matrix for the estimator of the
average partial effects from objects returned by \code{\link{apes}}.
}
\seealso{
\code{\link{apes}}
}
