% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in generics_summary.R
\name{summary.feglm}
\alias{summary.feglm}
\title{Summary method for fixed effects GLMs}
\usage{
\method{summary}{feglm}(object, ...)
}

\arguments{
\item{object}{'feglm' object}

\item{...}{additional arguments for S3 compliance (unused)}
}

\description{
Summary method for fixed effects GLMs
}

