% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in generics_summary.R
\name{summary.felm}
\alias{summary.felm}
\title{Summary method for fixed effects LMs}
\usage{
\method{summary}{felm}(object, type = "hessian", ...)
}

\arguments{
\item{object}{'felm' object}

\item{type}{'hessian' (no other type works with 'felm')}

\item{...}{additional arguments for S3 compliance (unused)}
}

\description{
Summary method for fixed effects LMs
}

