% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellQuadraticProgrammingAux.R
\name{top.genes}
\alias{top.genes}
\title{Top Variable Gene Identificaiton}
\usage{
top.genes(input.dir, output.dir, top.number.count = 500)
}
\arguments{
\item{input.dir}{The path to where the dataset is saved. Dataset should be saved as a tab-delimited table with row = gene, column = cell.}

\item{output.dir}{The path to where the output should be saved. The output will be saved as a tab-delimited table with NO row names or column names.}

\item{top.number.count}{The number of top variable genes to extract. Default to 500 genes}
}
\description{
This function identifies top n variable genes using method described in KeyGenes algorithm. In brief, this algorithm use cross validation based on LASSO regression. For detailed explanation, please refer to the paper listed in the note.
}
\note{
Code reference from Roost et. al., KeyGenes, a Tool to Probe Tissue Differentiation Using a Human Fetal Transcriptional Atlas, Cell Stem Cell Reports, 2015
}
\examples{
top.genes("~/Desktop/sample_data.txt", "~/Desktop/sample_gene_list_output.txt", top.number.count = 1000)
}
\keyword{extraction}
\keyword{gene}
\keyword{top}
\keyword{variable}
