% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransitionScoreCalculation.R
\name{transition.score}
\alias{transition.score}
\title{Transition Score Calculation}
\usage{
transition.score(multi.id.meta)
}
\arguments{
\item{multi.id.meta}{A data frame that contains cells with multiple identities. Column 1 - cell barcode, Column 2 - cell type call, Column 3 - counts, Column 4 - corresponding QP scores.}
}
\value{
A data frame that contains the calculated transition scores for each identity
}
\description{
This function calculates the transition scores for each cell state that is connected by cells with multiple identities.
}
\examples{
transition.score(multi.id.meta)
}
\keyword{Transition}
\keyword{metric}
