% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_vcov.R
\name{vcov.apes}
\alias{vcov.apes}
\title{Covariance matrix for APEs}
\usage{
\method{vcov}{apes}(object)
}
\arguments{
\item{object}{an object of class \code{"apes"}.}
}
\value{
A named matrix of covariance estimates.
}
\description{
Covariance matrix for the estimator of the
 average partial effects from objects returned by \code{\link{apes}}.
}
\seealso{
\code{\link{apes}}
}
