% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capybara-package.R
\docType{data}
\name{yotov2017}
\alias{yotov2017}
\title{Subset of the Trade Panel (1986-2006) from Yotov et al. (2017)}
\format{
\subsection{\code{yotov2017}}{

A data frame with 28,152 rows and 10 columns:
\describe{
\item{trade}{Nominal trade flows in current US dollars}
\item{log_dist}{Log of distance between the trading partners}
\item{cntg}{Indicator. Equal to 1 if country 'i' and 'j' share a common border}
\item{lang}{Indicator. Equal to 1 if country 'i' and 'j' speak the same official language}
\item{clny}{Indicator. Equal to 1 if country 'i' and 'j' share a colonial relationship}
\item{rta}{Indicator. Equal to 1 if country 'i' and 'j' are in a regional trade agreement}
\item{exp_year}{Exporter ISO country code and year}
\item{imp_year}{Importer ISO country code and year}
\item{pair}{Exporter and importer dyads}
\item{year}{Year of observation}
}
}
}
\source{
Advanced Guide to Trade Policy Analysis (ISBN: 978-92-870-4367-2)
}
\usage{
yotov2017
}
\description{
Aggregated exports at origin-destination-year level for 1986-2006
used for most of the examples in the package.
}
\keyword{datasets}
