% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.feglm}
\alias{autoplot.feglm}
\alias{autoplot.felm}
\title{Autoplot method for feglm objects}
\usage{
\method{autoplot}{feglm}(object, ...)

\method{autoplot}{felm}(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Additional arguments passed to the method. In this case,
the additional argument is \code{conf_level}, which is the confidence level for
the confidence interval.}
}
\value{
A ggplot object with the estimated coefficients and their confidence
intervals.

A ggplot object with the estimated coefficients and their confidence
intervals.
}
\description{
Extracts the estimated coefficients and their confidence
intervals.

Extracts the estimated coefficients and their confidence
}
\examples{
mod <- fepoisson(mpg ~ wt + drat | cyl, mtcars)
autoplot(mod, conf_level = 0.99)

mod <- felm(mpg ~ wt + drat | cyl, mtcars)
autoplot(mod, conf_level = 0.90)

}
