% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capybara-package.R
\docType{package}
\name{capybara-package}
\alias{capybara}
\alias{capybara-package}
\title{Generalized Linear Models (GLMs) with high-dimensional k-way fixed
effects}
\description{
Provides a routine to partial out factors with many levels during the optimization of the log-likelihood
function of the corresponding GLM. The package is based on the algorithm described in Stammann (2018). It also offers
an efficient algorithm to recover estimates of the fixed effects in a post-estimation routine and includes robust and
multi-way clustered standard errors. Further the package provides analytical bias corrections for binary choice
models derived by Fernandez-Val and Weidner (2016) and Hinz, Stammann, and Wanner (2020). This package is a ground up
rewrite with multiple refactors, optimizations, and new features compared to the original package \code{alpaca}. In its
current state, the package is stable and future changes will be limited to bug fixes and improvements, but not to
altering the functions' arguments or outputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pacha.dev/capybara/}
  \item \url{https://github.com/pachadotdev/capybara}
  \item Report bugs at \url{https://github.com/pachadotdev/capybara/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Other contributors:
\itemize{
  \item Joao Santos Silva [thesis advisor]
  \item Yoto Yotov [contributor]
}

}
