% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in capybara-package.R
\name{capybara-package}
\alias{capybara-package}
\title{Generalized Linear Models (GLMs) with high-dimensional k-way fixed
 effects}
\description{
Provides a routine to partial out factors with many levels during the optimization of the log-likelihood
function of the corresponding GLM. The package is based on the algorithm described in Stammann (2018). It also offers
an efficient algorithm to recover estimates of the fixed effects in a post-estimation routine and includes robust and
multi-way clustered standard errors. This package is a ground up rewrite with multiple refactors, optimizations, and
new features compared to the original package `alpaca`. In its current state, the package is stable and future changes
will be limited to bug fixes and improvements, but not to altering the functions' arguments or outputs.
}

