% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capybara-package.R
\docType{package}
\name{capybara-package}
\alias{capybara-package}
\title{Generalized Linear Models (GLMs) with high-dimensional k-way fixed
 effects}
\description{
Provides a routine to partial out factors with many levels during the
optimization of the log-likelihood function of the corresponding GLM. The
package is based on the algorithm described in Stammann (2018). It also
offers an efficient algorithm to recover estimates of the fixed effects in a
post-estimation routine and includes robust and multi-way clustered standard
errors. Further the package provides analytical bias corrections for binary
choice models derived by Fernández-Val and Weidner (2016) and Hinz, Stammann,
and Wanner (2020).
}
