% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in felm.R
\name{felm}
\alias{felm}
\title{LM fitting with high-dimensional k-way fixed effects}
\usage{
felm(formula = NULL, data = NULL, weights = NULL, vcov = NULL, control = NULL)
}

\arguments{
\item{formula}{an object of class \code{"formula"}: a symbolic description of the model to be fitted. \code{formula}
 must be of type \code{response ~ slopes | fixed_effects | cluster}.}

\item{data}{an object of class \code{"data.frame"} containing the variables in the model. The expected input is a
 dataset with the variables specified in \code{formula} and a number of rows at least equal to the number of variables
 in the model.}

\item{weights}{an optional string with the name of the prior weights variable in \code{data}.}

\item{vcov}{an optional character string specifying the type of variance-covariance estimator.
 One of \code{"iid"} (default OLS, ignore cluster part of formula), \code{"hetero"} (heteroskedastic-robust
 HC0, computed in C++ - no cluster variable needed), \code{"cluster"} (one-way sandwich using the cluster
 variable in the formula), \code{"m-estimator"} (M-estimator one-way sandwich), or \code{"dyadic"}
 (Cameron-Miller dyadic sandwich; requires two entity variables in the third part of the formula).
 When \code{NULL} (default), the type is inferred from the formula: if a cluster variable is present the
 standard sandwich is used, otherwise the inverse Hessian (IID) is returned.}

\item{control}{a named list of parameters for controlling the fitting process. See \link{fit_control} for details.}
}

\value{
A named list of class \code{"felm"}. The list contains the following
 eleven elements:
 \item{coefficients}{a named vector of the estimated coefficients}
 \item{fitted_values}{a vector of the estimated dependent variable}
 \item{weights}{a vector of the weights used in the estimation}
 \item{hessian}{a matrix with the numerical second derivatives}
 \item{null_deviance}{the null deviance of the model}
 \item{nobs}{a named vector with the number of observations used in the estimation indicating the dropped and
   perfectly predicted observations}
 \item{fe_levels}{a named vector with the number of levels in each fixed effect}
 \item{nms_fe}{a list with the names of the fixed effects variables}
 \item{formula}{the formula used in the model}
 \item{data}{the data used in the model after dropping non-contributing observations}
 \item{control}{the control list used in the model}
}

\description{
\link{feglm} can be used to fit linear models with many high-dimensional fixed effects. The estimation
 procedure is based on unconditional maximum likelihood and can be interpreted as a \dQuote{weighted demeaning}
 approach.
}

\examples{
ross2004_subset <- ross2004[ross2004$year == 1999, ]
ross2004_subset <- ross2004_subset[ross2004_subset$ltrade >
  quantile(ross2004_subset$ltrade, 0.75), ]

# Model with fixed effects
fit <- felm(ltrade ~ ldist | ctry1, ross2004_subset)
summary(fit)

ross2004_subset <- ross2004[ross2004$year \%in\% c(1994, 1999), ]
ross2004_subset <- ross2004_subset[ross2004_subset$ltrade >
  quantile(ross2004_subset$ltrade, 0.75), ]

# Model without fixed effects but with clustered standard errors
# Note: Use 0 to indicate no fixed effects when specifying clusters
fit <- felm(ltrade ~ ldist | 0 | year, ross2004_subset)
summary(fit)
}

