% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/felm.R
\name{felm}
\alias{felm}
\title{LM fitting with high-dimensional k-way fixed effects}
\usage{
felm(formula = NULL, data = NULL, weights = NULL)
}
\arguments{
\item{formula}{an object of class \code{"formula"}: a symbolic description of
the model to be fitted. \code{formula} must be of type \code{y ~ x | k},
where the second part of the formula refers to factors to be concentrated
out. It is also possible to pass additional variables to \code{\link{feglm}}
(e.g. to cluster standard errors). This can be done by specifying the third
part of the formula: \code{y ~ x | k | add}.}

\item{data}{an object of class \code{"data.frame"} containing the variables
in the model.}

\item{weights}{an optional string with the name of the 'prior weights'
variable in \code{data}.}
}
\value{
The function \code{\link{felm}} returns a named list of class
 \code{"felm"}.
}
\description{
A wrapper for \code{\link{feglm}} with
 \code{family = gaussian()}.
}
\examples{
mod <- felm(
  log(trade) ~ dist + lang + cntg + clny | exp_year + imp_year,
  trade_panel
)

summary(mod)
}
\references{
Gaure, S. (2013). "OLS with Multiple High Dimensional Category
 Variables". Computational Statistics and Data Analysis, 66.

Marschner, I. (2011). "glm2: Fitting generalized linear models
 with convergence problems". The R Journal, 3(2).

Stammann, A., F. Heiss, and D. McFadden (2016). "Estimating Fixed
 Effects Logit Models with Large Panel Data". Working paper.

Stammann, A. (2018). "Fast and Feasible Estimation of Generalized
 Linear Models with High-Dimensional k-Way Fixed Effects". ArXiv e-prints.
}
