% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in fenegbin.R
\name{fenegbin}
\alias{fenegbin}
\title{Negative Binomial model fitting with high-dimensional k-way fixed
 effects}
\usage{
fenegbin(
  formula = NULL,
  data = NULL,
  weights = NULL,
  beta_start = NULL,
  eta_start = NULL,
  init_theta = NULL,
  link = c("log", "identity", "sqrt"),
  offset = NULL,
  control = NULL
)
}

\arguments{
\item{init_theta}{an optional initial value for the theta parameter (see \link[MASS]{glm.nb}).}

\item{link}{the link function. Must be one of \code{"log"}, \code{"sqrt"}, or \code{"identity"}.}

\item{offset}{an optional formula or numeric vector specifying an a priori known component to be included in the
 linear predictor. If a formula, it should be of the form \code{~ variable}.}

\item{formula}{an object of class \code{"formula"}: a symbolic description of the model to be fitted. \code{formula}
 must be of type \code{response ~ slopes | fixed_effects | cluster}.}

\item{data}{an object of class \code{"data.frame"} containing the variables in the model. The expected input is a
 dataset with the variables specified in \code{formula} and a number of rows at least equal to the number of variables
 in the model.}

\item{weights}{an optional string with the name of the prior weights variable in \code{data}.}

\item{beta_start}{an optional vector of starting values for the structural parameters in the linear predictor.
 Default is \eqn{\boldsymbol{\beta} = \mathbf{0}}{\beta = 0}.}

\item{eta_start}{an optional vector of starting values for the linear predictor.}

\item{control}{a named list of parameters for controlling the fitting process. See \link{fit_control} for details.}
}

\value{
A named list of class \code{"feglm"}. The list contains the following
 eighteen elements:
 \item{coefficients}{a named vector of the estimated coefficients}
 \item{eta}{a vector of the linear predictor}
 \item{weights}{a vector of the weights used in the estimation}
 \item{hessian}{a matrix with the numerical second derivatives}
 \item{deviance}{the deviance of the model}
 \item{null_deviance}{the null deviance of the model}
 \item{conv}{a logical indicating whether the model converged}
 \item{iter}{the number of iterations needed to converge}
 \item{theta}{the estimated theta parameter}
 \item{iter_outer}{the number of outer iterations}
 \item{conv_outer}{a logical indicating whether the outer loop converged}
 \item{nobs}{a named vector with the number of observations used in the estimation indicating the dropped and
  perfectly predicted observations}
 \item{fe_levels}{a named vector with the number of levels in each fixed effects}
 \item{nms_fe}{a list with the names of the fixed effects variables}
 \item{formula}{the formula used in the model}
 \item{data}{the data used in the model after dropping non-contributing observations}
 \item{family}{the family used in the model}
 \item{control}{the control list used in the model}
}

\description{
A routine that uses the same internals as \link{feglm}.
}

\examples{
# check the felm examples for the details about clustered standard errors

ross2004_subset <- ross2004[ross2004$year == 1999, ]
ross2004_subset <- ross2004_subset[ross2004_subset$ltrade >
  quantile(ross2004_subset$ltrade, 0.75), ]

fit <- fenegbin(ltrade ~ ldist | ctry1, ross2004_subset)

summary(fit)
}

