% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fenegbin.R
\name{fenegbin}
\alias{fenegbin}
\title{Negative Binomial model fitting with high-dimensional k-way fixed
effects}
\usage{
fenegbin(
  formula = NULL,
  data = NULL,
  weights = NULL,
  beta_start = NULL,
  eta_start = NULL,
  init_theta = NULL,
  link = c("log", "identity", "sqrt"),
  control = NULL
)
}
\arguments{
\item{formula}{an object of class \code{"formula"}: a symbolic description of
the model to be fitted. \code{formula} must be of type \code{y ~ x | k},
where the second part of the formula refers to factors to be concentrated
out. It is also possible to pass clustering variables to \code{\link{feglm}}
as \code{y ~ x | k | c}.}

\item{data}{an object of class \code{"data.frame"} containing the variables
in the model.}

\item{weights}{an optional string with the name of the 'prior weights'
variable in \code{data}.}

\item{beta_start}{an optional vector of starting values for the structural
parameters in the linear predictor. Default is
\eqn{\boldsymbol{\beta} = \mathbf{0}}{\beta = 0}.}

\item{eta_start}{an optional vector of starting values for the linear
predictor.}

\item{init_theta}{an optional initial value for the theta parameter (see
\code{\link[MASS]{glm.nb}}).}

\item{link}{the link function. Must be one of \code{"log"}, \code{"sqrt"}, or
\code{"identity"}.}

\item{control}{a named list of parameters for controlling the fitting
process. See \code{\link{feglm_control}} for details.}
}
\value{
A named list of class \code{"feglm"}.
}
\description{
A routine that uses the same internals as \code{\link{feglm}}.
}
\examples{
# same as the example in fepoisson but with overdispersion/underdispersion
mod <- fenegbin(
  trade ~ log_dist + lang + cntg + clny | exp_year + imp_year,
  trade_panel
)

summary(mod)

}
