% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fepoisson.R
\name{fepoisson}
\alias{fepoisson}
\title{Poisson model fitting high-dimensional with k-way fixed effects}
\usage{
fepoisson(
  formula = NULL,
  data = NULL,
  weights = NULL,
  beta.start = NULL,
  eta.start = NULL,
  control = NULL
)
}
\arguments{
\item{formula}{an object of class \code{"formula"}: a symbolic description of
the model to be fitted. \code{formula} must be of type \code{y ~ x | k},
where the second part of the formula refers to factors to be concentrated
out. It is also possible to pass additional variables to \code{\link{feglm}}
(e.g. to cluster standard errors). This can be done by specifying the third
part of the formula: \code{y ~ x | k | add}.}

\item{data}{an object of class \code{"data.frame"} containing the variables
in the model.}

\item{weights}{an optional string with the name of the 'prior weights'
variable in \code{data}.}

\item{beta.start}{an optional vector of starting values for the structural
parameters in the linear predictor. Default is
\eqn{\boldsymbol{\beta} = \mathbf{0}}{\beta = 0}.}

\item{eta.start}{an optional vector of starting values for the linear
predictor.}

\item{control}{a named list of parameters for controlling the fitting
process. See \code{\link{feglm_control}} for details.}
}
\description{
A wrapper for \code{\link{feglm}} with
 \code{family = poisson()}.
}
\examples{
# same as the example in feglm but with less typing
mod <- fepoisson(
  trade ~ dist + lang + cntg + clny | exp_year + imp_year,
  trade_panel
)

summary(mod)
}
