% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_control.R
\name{fit_control}
\alias{fit_control}
\title{Set \code{feglm} Control Parameters}
\usage{
fit_control(
  dev_tol = 1e-06,
  center_tol = 1e-06,
  collin_tol = 1e-07,
  step_halving_factor = 0.5,
  alpha_tol = 1e-06,
  iter_max = 25L,
  iter_center_max = 10000L,
  iter_inner_max = 50L,
  iter_alpha_max = 10000L,
  iter_interrupt = 1000L,
  iter_ssr = 10L,
  keep_tx = FALSE
)
}
\arguments{
\item{dev_tol}{tolerance level for the first stopping condition of the
maximization routine. The stopping condition is based on the relative change
of the deviance in iteration \eqn{r} and can be expressed as follows:
\eqn{|dev_{r} - dev_{r - 1}| / (0.1 + |dev_{r}|) < tol}{|dev - devold| /
 (0.1 + |dev|) < tol}. The default is \code{1.0e-08}.}

\item{center_tol}{tolerance level for the stopping condition of the centering
algorithm. The stopping condition is based on the relative change of the
centered variable similar to the \code{'lfe'} package. The default is
\code{1.0e-08}.}

\item{collin_tol}{tolerance level for detecting collinearity. The default is
\code{1.0e-07}.}

\item{step_halving_factor}{numeric indicating the factor by which the step
size is halved to iterate towards convergence. This is used to control the
step size during optimization. The default is \code{0.5}.}

\item{alpha_tol}{tolerance for fixed effects (alpha) convergence.
The default is \code{1.0e-06}.}

\item{iter_max}{unsigned integer indicating the maximum number of iterations
in the maximization routine. The default is \code{25L}.}

\item{iter_center_max}{unsigned integer indicating the maximum number of
iterations in the centering algorithm. The default is \code{10000L}.}

\item{iter_inner_max}{unsigned integer indicating the maximum number of
iterations in the inner loop of the centering algorithm. The default is
\code{50L}.}

\item{iter_alpha_max}{maximum iterations for fixed effects computation.
The default is \code{10000L}.}

\item{iter_interrupt}{unsigned integer indicating the maximum number of
iterations before the algorithm is interrupted. The default is \code{1000L}.}

\item{iter_ssr}{unsigned integer indicating the number of iterations
to skip before checking if the sum of squared residuals improves. The default
is \code{10L}.}

\item{keep_tx}{logical indicating if the centered regressor matrix should be
stored. The centered regressor matrix is required for some covariance
estimators, bias corrections, and average partial effects. This option saves
some computation time at the cost of memory. The default is \code{TRUE}.}
}
\value{
A named list of control parameters.
}
\description{
Set and change parameters used for fitting \code{\link{feglm}}.
Termination conditions are similar to \code{\link[stats]{glm}}.
}
\examples{
fit_control(0.05, 0.05, 10L, 10L, TRUE, TRUE, TRUE)

}
\seealso{
\code{\link{feglm}}
}
