% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_helpers.R
\name{fit_control}
\alias{fit_control}
\title{Set \code{feglm} and \code{felm} Control Parameters}
\usage{
fit_control(
  dev_tol = 1e-08,
  center_tol = 1e-08,
  collin_tol = 1e-07,
  iter_max = 25L,
  iter_center_max = 10000L,
  iter_inner_max = 50L,
  iter_interrupt = 1000L,
  iter_ssr = 10L,
  direct_qr_threshold = 0.9,
  qr_collin_tol_multiplier = 1,
  chol_stability_threshold = 1e-12,
  safe_division_min = 1e-12,
  safe_log_min = 1e-12,
  newton_raphson_tol = 1e-08,
  demean_extra_projections = 0L,
  demean_warmup_iterations = 15L,
  demean_projections_after_acc = 5L,
  demean_grand_acc_frequency = 20L,
  demean_ssr_check_frequency = 40L,
  irons_tuck_eps = 1e-14,
  alpha_convergence_tol = 1e-08,
  alpha_iter_max = 10000L,
  limit = 10L,
  trace = FALSE,
  drop_pc = TRUE,
  keep_mx = FALSE,
  use_weights = TRUE
)
}
\arguments{
\item{dev_tol}{tolerance level for the first stopping condition of the
maximization routine. The stopping condition is based on the relative change
of the deviance in iteration \eqn{r} and can be expressed as follows:
\eqn{|dev_{r} - dev_{r - 1}| / (0.1 + |dev_{r}|) < tol}{|dev - devold| /
 (0.1 + |dev|) < tol}. The default is \code{1.0e-08}.}

\item{center_tol}{tolerance level for the stopping condition of the centering
algorithm. The default is \code{1.0e-08}.}

\item{collin_tol}{tolerance level for detecting collinearity. The default is
\code{1.0e-10}.}

\item{iter_center_max}{unsigned integer indicating the maximum number of
iterations in the centering algorithm. The default is \code{10000L}.}

\item{keep_mx}{logical indicating if the centered regressor matrix should be
stored. The centered regressor matrix is required for some covariance
estimators, bias corrections, and average partial effects. This option saves
some computation time at the cost of memory. The default is \code{TRUE}.}

\item{use_weights}{logical indicating whether to use weights. If \code{FALSE},
weights are ignored for performance. The default is \code{TRUE}.}
}
\value{
A named list of control parameters.
}
\description{
Set and change parameters used for fitting \code{\link{feglm}}
and \code{felm}.Termination conditions are similar to
\code{\link[stats]{glm}}.
}
\examples{
fit_control(0.05, 0.05, 10L, 10L, TRUE, TRUE, TRUE)

}
\seealso{
\code{\link{feglm}}
}
