% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_effects.R
\name{fixed_effects}
\alias{fixed_effects}
\title{Recover the estimates of the fixed effects after fitting (G)LMs}
\usage{
fixed_effects(object = NULL, alpha_tol = 1e-08)
}
\arguments{
\item{object}{an object of class \code{"feglm"}.}

\item{alpha_tol}{tolerance level for the stopping condition. The algorithm is
stopped at iteration \eqn{i} if \eqn{||\boldsymbol{\alpha}_{i} -
 \boldsymbol{\alpha}_{i - 1}||_{2} < tol ||\boldsymbol{\alpha}_{i - 1}||
 {2}}{||\Delta \alpha|| < tol ||\alpha_old||}. Default is \code{1.0e-08}.}
}
\value{
A named list containing named vectors of estimated fixed effects.
}
\description{
The system might not have a unique solution since we do not take
collinearity into account. If the solution is not unique, an estimable
function has to be applied to our solution to get meaningful estimates of
the fixed effects.
}
\examples{
# check the feglm examples for the details about clustered standard errors

# subset trade flows to avoid fitting time warnings during check
set.seed(123)
trade_2006 <- trade_panel[trade_panel$year == 2006, ]
trade_2006 <- trade_2006[sample(nrow(trade_2006), 1000), ]

mod <- fepoisson(
  trade ~ log_dist + lang + cntg + clny | exp_year + imp_year,
  trade_2006
)

fixed_effects(mod)

}
\references{
Stammann, A. (2018). "Fast and Feasible Estimation of Generalized
Linear Models with High-Dimensional k-way Fixed Effects". ArXiv e-prints.

Gaure, S. (n. d.). "Multicollinearity, identification, and
estimable functions". Unpublished.
}
\seealso{
\code{\link{felm}}, \code{\link{feglm}}
}
