% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capybara-package.R
\docType{data}
\name{ross2004}
\alias{ross2004}
\title{Subset of WTO data from Ross (2004)}
\format{
\subsection{\code{ross2004}}{

A data frame with 234,597 rows and 22 columns:
\describe{
\item{ltrade}{Log of bilateral trade between i and j at time t}
\item{bothin}{Binary variable which is unity if both i and j are GATT/WTO members at t}
\item{onein}{Binary variable which is unity if either i or j is a GATT/WTO member at t}
\item{gsp}{Binary variable which is unity if i was a GSP beneficiary of j or vice versa at t}
\item{ldist}{Log of distance between i and j}
\item{lrgdp}{Log of real GDP}
\item{lrgdppc}{Log of real GDP per capita}
\item{regional}{Binary variable which is unity if i and j are in the same region}
\item{custrict}{Binary variable which is unity if i and j are in the same customs union}
\item{comlang}{Binary variable which is unity if i and j share a common language}
\item{border}{Binary variable which is unity if i and j share a border}
\item{landl}{Number of landlocked countries in the country-pair (0, 1, or 2)}
\item{island}{Number of island nations in the pair (0, 1, or 2)}
\item{lareap}{Log of the area of the country (in square kilometers)}
\item{comcol}{Binary variable which is unity if i and j were ever colonies after 1945 with the same colonizer}
\item{curcol}{Binary variable which is unity if i and j are colonies at time t}
\item{colony}{Binary variable which is unity if i ever colonized j or vice versa}
\item{comctry}{Binary variable which is unity if i and j remained part of the same nation during the sample (e.g., France and Guadeloupe)}
\item{ctry1}{Country 1 ISO-3 code}
\item{ctry2}{Country 2 ISO-3 code}
\item{pair}{Country pair undirected dyads}
\item{year}{Year of observation}
}
}
}
\source{
Do We Really Know That the WTO Increases Trade? (DOI: 10.1257/000282804322970724)
}
\usage{
ross2004
}
\description{
Data from Ross (2004) used to show the different variance-covariance estimators from
Cameron and Miller (2014).
}
\keyword{datasets}
