% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_vcov.R
\name{vcov.felm}
\alias{vcov.felm}
\title{Covariance matrix for LMs}
\usage{
\method{vcov}{felm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"felm"}.}

\item{...}{additional arguments (currently ignored).}
}
\value{
A named matrix of covariance estimates.
}
\description{
Covariance matrix for the estimator of the structural parameters from objects returned by \link{felm}.
The covariance is computed during model fitting - either the inverse Hessian (default) or the sandwich estimator if
cluster variables are specified in the formula.
}
\examples{
# Model with clustering - returns sandwich covariance
yotov2017_subset <- yotov2017[yotov2017$year == 2006, ]
yotov2017_subset <- yotov2017_subset[yotov2017_subset$trade > 0, ]
mod_cl <- felm(trade ~ log_dist | exp_year | year, yotov2017_subset)
vcov(mod_cl)

}
\references{
Cameron, C., J. Gelbach, and D. Miller (2011). "Robust Inference With Multiway Clustering". Journal of
Business & Economic Statistics 29(2).
}
\seealso{
\link{felm}
}
