% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in generics_vcov.R
\name{vcov.felm}
\alias{vcov.felm}
\title{Covariance matrix for LMs}
\usage{
\method{vcov}{felm}(object, ...)
}

\arguments{
\item{object}{an object of class \code{"felm"}.}

\item{...}{additional arguments (currently ignored).}
}

\value{
A named matrix of covariance estimates.
}

\description{
Covariance matrix for the estimator of the structural parameters from objects returned by \link{felm}.
 The covariance is computed during model fitting - either the inverse Hessian (default) or the sandwich estimator if
 cluster variables are specified in the formula.
}

\seealso{
\link{felm}
}

\examples{
# Model with clustering - returns sandwich covariance
ross2004_s1 <- ross2004[ross2004$year == 1994, ]
ross2004_s1 <- ross2004_s1[ross2004_s1$ltrade >
  quantile(ross2004_s1$ltrade, 0.75), ]

ross2004_s2 <- ross2004[ross2004$year == 1999, ]
ross2004_s2 <- ross2004_s2[ross2004_s2$ltrade >
  quantile(ross2004_s2$ltrade, 0.75), ]

ross2004_subset <- rbind(ross2004_s1, ross2004_s2)

fit <- felm(ltrade ~ ldist | ctry1 | year, ross2004_subset)

vcov(fit)
}

